Eliminate Logger_Impl
